feat: add DTO classes and update return types in service and controller
- Added new DTOs for better data structuring.
- Updated service and controller methods to use DTOs instead of entities.
- Changed method return types to reflect the new DTO classes.

+++ src/main/java/br/com/empiricus/EmpiricusADM/Service/UserService.java
@@ -3,31 +3,42 @@ package br.com.empiricus.EmpiricusADM.Service;
 import br.com.empiricus.EmpiricusADM.Model.User;
 import br.com.empiricus.EmpiricusADM.Repository.UserRepository;
 import br.com.empiricus.EmpiricusADM.Service.Exceptions.ItemNotFound;
+import br.com.empiricus.EmpiricusADM.Service.Exceptions.MismatchId;
+import br.com.empiricus.EmpiricusADM.dto.UserDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository repo;
-    private final MailService mailService;
 
-    public List<User> findAllUsers(){
-        return repo.findAll();
+
+    public List<UserDTO> findAllUsers() {
+        return repo.findAll().stream().map(UserDTO::new).collect(Collectors.toList());
     }
-    public User findUserById(Long id){
-        return repo.findById(id)
-                .orElseThrow(()->new ItemNotFound("User exception id: "+id));
+
+    public UserDTO findUserById(Long id) {
+        User user = repo.findById(id)
+                .orElseThrow(() -> new ItemNotFound("User exception id: " + id));
+        return new UserDTO(user);
     }
-    public User createUser(User user){
-        return  repo.save(user);
+
+    public UserDTO createUser(User user) {
+        return new UserDTO(repo.save(user));
     }
-    public User UpdateUser(Long id, User user){
-    User item = repo.getReferenceById(id);
-    update(item, user);
-    return repo.save(item);
+
+    public UserDTO updateUser(Long id, User user) {
+        if(!Objects.equals(id, user.getId())){
+            throw new MismatchId("Id passed on path "+ id+" isn't equal to id passed onto the object "+ user.getId());
+        }
+        User item = repo.getReferenceById(id);
+        update(item, user);
+        return new UserDTO(repo.save(item));
     }
 
     private void update(User item, User user) {
@@ -36,8 +47,14 @@ public class UserService {
         item.setData_atualizacao(user.getData_atualizacao());
         item.setEh_admin(user.isEh_admin());
     }
-    public void DeleteUser(Long id){
+
+    public void deleteUser(Long id) {
+        if(findUserById(id)!=null){
+
         repo.deleteById(id);
+        } else {
+            throw new ItemNotFound("Item id: "+id+" was not found");
+        }
     }
